requirements.txt
lightrag-hku
openai
aiohttp
traceloop-sdk
firecrawl-py
pypdf
opentelemetry-exporter-otlp
